refactoring

+++ src/main/java/com/taskmanagement/task_management_system/Service/TeamService.java
@@ -7,6 +7,7 @@ import com.taskmanagement.task_management_system.Enum.UserRole;
 import com.taskmanagement.task_management_system.Exception.Resource.ResourceAlreadyExistException;
 import com.taskmanagement.task_management_system.Exception.Resource.ResourceNotFoundException;
 import com.taskmanagement.task_management_system.Mapper.TeamMapper;
+import com.taskmanagement.task_management_system.Model.CustomUserDetails;
 import com.taskmanagement.task_management_system.Model.dto.team.*;
 import com.taskmanagement.task_management_system.Model.entity.PendingJoiningTeam;
 import com.taskmanagement.task_management_system.Model.entity.Team;
@@ -14,15 +15,12 @@ import com.taskmanagement.task_management_system.Model.entity.Users;
 import com.taskmanagement.task_management_system.Repository.PendingTeamRepository;
 import com.taskmanagement.task_management_system.Repository.TeamRepository;
 import com.taskmanagement.task_management_system.Repository.UserRepository;
-import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.repository.query.Param;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
-
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
@@ -53,20 +51,18 @@ public class TeamService extends BaseService<Team, Long> {
         );
     }
 
-    public TeamInfo saveTeam(TeamRequest request , Long userId) {
+    public TeamInfo createTeam(TeamRequest request , Long userId) {
         Team team = teamMapper.toEntity(request);
         Users user = userService.getUserEntity(userId);
         team.addUser(user);
-        TeamInfo savedTeam= teamMapper.toDto(super.save(team));
-
-        return savedTeam;
+        return teamMapper.toDto(super.save(team));
     }
 
     public TeamInfo updateTeam(Long id , UpdateTeamRequest dto , String createdBy) {
-        Team team = teamRepository.findByCreatedBy(id , createdBy).orElseThrow(
+        Team team = teamRepository.findById(id).orElseThrow(
                 ()-> new ResourceNotFoundException("team not found with id: "+ id)
         );
-
+        verifyOwnerOrThrow(team , createdBy);
         teamMapper.updateTeam(dto , team);
         Team updatedTeam = teamRepository.save(team);
         return teamMapper.toDto(updatedTeam);
@@ -74,10 +70,10 @@ public class TeamService extends BaseService<Team, Long> {
     }
 
     public TeamWithMembers getMembers(Long id , String createdBy) {
-        Team team = teamRepository.findByCreatedBy(id , createdBy).orElseThrow(
+        Team team = teamRepository.findById(id).orElseThrow(
                 ()-> new ResourceNotFoundException("team not found with id: " + id)
         );
-
+        verifyOwnerOrThrow(team , createdBy);
         return teamMapper.membersDto(team);
     }
 
@@ -92,17 +88,24 @@ public class TeamService extends BaseService<Team, Long> {
     @Transactional
     public void addMembers(Long id, List<Long> userIds , String createdBy) {
 
-        Team team = teamRepository.findByCreatedBy(id , createdBy).orElseThrow(
+        Team team = teamRepository.findById(id).orElseThrow(
                 ()-> new ResourceNotFoundException("team not found with id: " + id)
         );
+        verifyOwnerOrThrow(team , createdBy);
 
         List<Users> users = userRepository.findAllById(userIds);
+        boolean isMember ;
 
         if(users.size() != userIds.size()) {
             throw new ResourceNotFoundException("Some users were not found");
         }
 
         for (Users user : users) {
+          isMember= teamRepository.existsByUsersIdAndId(user.getId() , id);
+            if(isMember) {
+                throw new ResourceAlreadyExistException("user with id " + user.getId() + " is already a member of team " + id);
+            }
+
             user.addTeam(team);
         }
     }
@@ -113,9 +116,11 @@ public class TeamService extends BaseService<Team, Long> {
                         new ResourceNotFoundException(
                                 "User not found with id: " + userId));
 
-        Team team = teamRepository.findByCreatedBy(id,createdBy).orElseThrow(
+        Team team = teamRepository.findById(id).orElseThrow(
                 ()-> new ResourceNotFoundException("team not found with id: " + id)
         );
+        verifyOwnerOrThrow(team , createdBy);
+
         boolean isMember = teamRepository.existsByUsersIdAndId(userId , id);
         if(isMember) {
             throw new ResourceAlreadyExistException("user with id " + userId + " is already a member of team " + id);
@@ -126,10 +131,11 @@ public class TeamService extends BaseService<Team, Long> {
 
     public void removeMemberFromTeam(Long teamId, Long userId , String createdBy) {
 
-        Team team = teamRepository.findByCreatedBy(teamId, createdBy).orElseThrow(
+        Team team = teamRepository.findById(teamId).orElseThrow(
                 ()-> new ResourceNotFoundException("team not found with id: " + teamId)
         );
 
+        verifyOwnerOrThrow(team , createdBy);
 
         Users user = userRepository.findById(userId)
                 .orElseThrow(() ->
@@ -167,19 +173,20 @@ public class TeamService extends BaseService<Team, Long> {
     }
 
     public List<TeamInfo> search(String teamName) {
-      List<TeamInfo> team = teamRepository.findTeamInfoByName(teamName);
-      return team;
+      return teamRepository.findTeamInfoByName(teamName);
 
     }
     public TeamAvailableUsers getAvailableUsers(Long teamId , Long userId) {
 
         String createdBy = userService.getUserEntity(userId).getUsername();
 
-        Team team = teamRepository.findByCreatedBy(teamId , createdBy).orElseThrow(
+        Team team = teamRepository.findById(teamId).orElseThrow(
                 ()-> new ResourceNotFoundException(
                         "team not found with id: "+ teamId
                 )
         );
+
+        verifyOwnerOrThrow(team , createdBy);
 
         List<Long> userIds = teamRepository.findAvailableUsers(teamId);
 
@@ -194,9 +201,10 @@ public class TeamService extends BaseService<Team, Long> {
                     "User not found with id: " + userId
             );
         }
-        Team team = teamRepository.findByCreatedBy(teamId , createdBy).orElseThrow(
+        Team team = teamRepository.findById(teamId).orElseThrow(
                 ()-> new ResourceNotFoundException( "Team not found with id: " + teamId)
         );
+        verifyOwnerOrThrow(team , createdBy);
 
         return teamRepository.existsByUsersIdAndId(userId , team.getId());
 
@@ -213,9 +221,12 @@ public class TeamService extends BaseService<Team, Long> {
 
     public TeamWithMembers replaceMembers(Long teamId, AddUsersToTeamRequest request , String createdBy) {
 
-        Team team = teamRepository.findByCreatedBy(teamId,createdBy).orElseThrow(
+        Team team = teamRepository.findById(teamId).orElseThrow(
                 ()-> new ResourceNotFoundException("team not found with id: " + teamId)
         );
+
+        verifyOwnerOrThrow(team , createdBy);
+
         List<Users> users = userRepository.findAllById(request.getUserIds());
         if (users.size() != request.getUserIds().size()) {
             throw new ResourceNotFoundException("Some users not found");
@@ -266,10 +277,11 @@ public class TeamService extends BaseService<Team, Long> {
                 .orElseThrow(
                         ()->new ResourceNotFoundException("user with id: " + pending.getUserId() + " not found")
                 );
-        Team team = teamRepository.findByCreatedBy(pending.getTeamId() , createdBy).orElseThrow(
+        Team team = teamRepository.findById(pending.getTeamId()).orElseThrow(
                 ()->new ResourceNotFoundException("team with id: " + pending.getTeamId() + " not found")
 
         );
+        verifyOwnerOrThrow(team , createdBy);
         team.addUser(user);
         pendingTeamRepository.delete(pending);
     }
@@ -287,5 +299,26 @@ public class TeamService extends BaseService<Team, Long> {
     public Long countTeams(Long userId){
         return teamRepository.countByUsersId(userId);
     }
+
+
+    ///  validations
+
+    ///  create teams -> admin - manager
+    public void verifyOwnerOrThrow(Team team , String username) {
+        if(!team.getCreatedBy().equals(username)) {
+            throw new AccessDeniedException("team not found with id: " + team.getId());
+        }
+
+    }
+
+    ///  get role
+     public void verifyCurrentCanCreate(CustomUserDetails currentUser) {
+         Users user = currentUser.user();
+         UserRole role = user.getRole();
+         if(role == UserRole.USER) {
+             throw new AccessDeniedException("you don't have permissions to perform this action.");
+         }
+
+     }
 
 }
